// src copy/pages/LogoutPage/types.tsx
import { SystemStyleObject } from "@styled-system/css";

export type LogoutPageProps = {
    styleConfig: {
        rootStyles: SystemStyleObject;
        titleBox: SystemStyleObject;
        pageTitle: SystemStyleObject;
        pageContent: SystemStyleObject;
        content: SystemStyleObject;
        link: SystemStyleObject;
    };
};